Modified two functions
- Modified CreateCourse()
   Adds the role of creator to a user
- Modified DeleteCourse()
   Checks if 10 Minutes have passed
   since creation

+++ course/course.go
@@ -13,7 +13,8 @@ import (
 )
 
 //CreateCourse takes a name,enrollkey and description and adds a course and forum with that Name in the Database
-func CreateCourse(name, enrollkey string, description null.String) {
+func CreateCourse(name, enrollkey string, description null.String, usersid []int) {
+	//Connets to the database
 	db, err := sql.Open("mysql", "root:root@tcp(127.0.0.1:3306)/learningbay24")
 
 	if err != nil {
@@ -21,17 +22,30 @@ func CreateCourse(name, enrollkey string, description null.String) {
 	} else {
 		fmt.Println("Connected")
 	}
+
+	//Creates a Forum struct (Forum has to be created first because of Foreign Key)
 	f := &models.Forum{Name: name}
+	//Inserts into database
 	err = f.Insert(context.Background(), db, boil.Infer())
 	if err != nil {
 		panic(err.Error())
 	}
+
+	//Creates a Course struct
 	c := &models.Course{Name: name, ForumID: f.ID, Description: description}
+	//Inserts into database
 	err = c.Insert(context.Background(), db, boil.Infer())
 	if err != nil {
 		panic(err.Error())
+	} else {
+		shasc := models.UserHasCourse{UserID: usersid[0], CourseID: c.ID, RoleID: 1}
+		err = shasc.Insert(context.Background(), db, boil.Infer())
+		if err != nil {
+			DeleteCourse(c.ID)
+			panic(err.Error())
+		}
 	}
-	db.Close()
+	defer db.Close()
 }
 
 //UpdateCourse takes the ID of a existing course and overwrites the corespoding course and forum with the new Strings(name,enrollkey and description)
@@ -71,7 +85,7 @@ func UpdateCourse(id int, name, enrollkey string, description null.String) {
 		panic(err.Error())
 	}
 
-	db.Close()
+	defer db.Close()
 }
 
 //DeleteCourse takes a ID and deletes the course and the forum associated with it
@@ -88,24 +102,31 @@ func DeleteCourse(id int) {
 		panic(err.Error())
 	}
 
-	cdel, err := c.Delete(context.Background(), db, true)
-	fmt.Println(cdel)
-	if err != nil {
-		panic(err.Error())
-	}
+	//Checks if more than 10 Minutes have passed wont delete if thats the case
+	curTime := time.Now()
+	diff := curTime.Sub(c.CreatedAt.Time)
+	if diff.Minutes() < 10 {
+		fmt.Println("Only ", diff.Minutes(), "Have passed")
+		cdel, err := c.Delete(context.Background(), db, true)
+		fmt.Println(cdel)
+		if err != nil {
+			panic(err.Error())
+		}
 
-	f, err := models.FindForum(context.Background(), db, id)
-	if err != nil {
-		panic(err.Error())
-	}
+		f, err := models.FindForum(context.Background(), db, id)
+		if err != nil {
+			panic(err.Error())
+		}
 
-	fdel, err := f.Delete(context.Background(), db, true)
-	fmt.Println(fdel)
-	if err != nil {
-		panic(err.Error())
+		fdel, err := f.Delete(context.Background(), db, true)
+		fmt.Println(fdel)
+		if err != nil {
+			panic(err.Error())
+		}
+	} else {
+		fmt.Println("More than 10 Minutes have passed")
 	}
-
-	db.Close()
+	defer db.Close()
 }
 
 //DeactivateCourse takes a ID and deactivates the course and the forum associated with it
@@ -139,5 +160,5 @@ func DeactivateCourse(id int) {
 		panic(err.Error())
 	}
 
-	db.Close()
+	defer db.Close()
 }
